+++ src/Params/OptionsInterface.php
@@ -2,39 +2,24 @@
 
 namespace Params;
 
+use Params\Immutable\ImmutableOptionsInterface;
+
 /**
  * Class that wraps an associative array for
  * more convenient access of keys and values.
  */
-interface OptionsInterface
+interface OptionsInterface extends ImmutableOptionsInterface
 {
-    /**
-     * Returns whether the key exists or not.
-     *
-     * @param string $key name of key to check
-     *
-     * @return bool true, if key exists; false otherwise
-     */
-    public function has($key);
-
-    /**
-     * Returns the value for the given key.
-     *
-     * @param string $key name of key
-     * @param mixed $default value to return if key doesn't exist
-     *
-     * @return mixed value for that key or default given
-     */
-    public function get($key, $default = null);
-
     /**
      * Sets a given value for the specified key.
      *
-     * @param string $key name of entry
+     * @param mixed $key name of entry
      * @param mixed $value value to set for the given key
      * @param bool $replace whether or not to replace values of existing keys
      *
      * @return Options self instance for fluent API
+     *
+     * @throws \InvalidArgumentException on empty string or null key given
      */
     public function set($key, $value, $replace = true);
 
@@ -51,7 +36,7 @@ interface OptionsInterface
     /**
      * Removes the given key from the internal array.
      *
-     * @param string $key name of key to remove
+     * @param mixed $key name of key to remove
      *
      * @return Options self instance for fluent API
      */
@@ -68,50 +53,9 @@ interface OptionsInterface
      * Runs given callback for every key on the current level. The callback
      * should accept key and value as parameters and return the new value.
      *
-     * @param callback $callback callback to run for each entry of the current level
+     * @param mixed $callback callback to run for each entry of the current level
      *
      * @return Options self instance for fluent API
      */
     public function map($callback);
-
-    /**
-     * Returns all first level key names.
-     *
-     * @return array of keys
-     */
-    public function getKeys();
-
-    /**
-     * Allows to search for specific data values via JMESPath expressions.
-     *
-     * Some example expressions as a quick start:
-     *
-     * - "nested.key"           returns the value of the nested "key"
-     * - "nested.*"             returns all values available under the "nested" key
-     * - "*.key"                returns all values of "key"s on any second level array
-     * - "[key, nested.key]"    returns first level "key" value and the first value of the "nested" key array
-     * - "[key, nested[0]"      returns first level "key" value and the first value of the "nested" key array
-     * - "nested.key || key"    returns the value of the first matching expression
-     *
-     * @see http://jmespath.readthedocs.org/en/latest/
-     * @see https://github.com/mtdowling/jmespath.php
-     *
-     * @param string $expression JMESPath expression to evaluate on stored data
-     *
-     * @return mixed|null data in various types (scalar, array etc.) depending on the found results
-     *
-     * @throws \JmesPath\SyntaxErrorException on invalid expression syntax
-     * @throws \RuntimeException e.g. if JMESPath cache directory cannot be written
-     * @throws \InvalidArgumentException e.g. if JMESPath builtin functions can't be called
-     */
-    public function getValues($expression = '*');
-
-    /**
-     * Returns the data as an associative array.
-     *
-     * @param bool $recursive whether or not nested arrays should be included as array or object
-     *
-     * @return array of all data
-     */
-    public function toArray($recursive = true);
 }