+++ tests/Params/Tests/ImmutableParametersTraitTester.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Params\Tests;
-
-class ImuutableParametersTraitTester
-{
-    use \Params\ImmutableParametersTrait;
-}