+++ tests/Params/Tests/OptionsTraitTester.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Params\Tests;
-
-class OptionsTraitTester
-{
-    use \Params\OptionsTrait;
-}