+++ tests/Params/Tests/ParametersTraitTester.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Params\Tests;
-
-class ParametersTraitTester
-{
-    use \Params\ParametersTrait;
-}