+++ tests/Params/Tests/TraitTester.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Params\Tests;
+
+class TraitTester
+{
+    use \Params\ParametersTrait;
+    use \Params\OptionsTrait;
+    //use \Params\SettingsTrait;
+}